E3: format family extinct -- warnings 308 -> 272
All 36 remaining format=/format-overflow sites.  No live bug on retail
data, but each overflow was latent memory corruption on modded data:
struct-field sprintf targets get precision bounds (%.*s with
MAX_FILENAME_LEN/NAME_LENGTH arithmetic -- explosion LOD filenames,
bmpman frame suffixes, wing-wave name bashing, ship_create names,
debrief voice files), tight locals get sized to their real worst case
(campaign savefile names, medals label/filenames, sexp error text,
gr_print_timestamp's exact-fit h/m/s), the '0'-flag-with-precision
family respells %0.2d as the byte-identical %.2d, and show_framerate's
fix-fed %d gets an int cast.

Delta: 36 removed, 0 added.  Tests green, boot renders 73 frames.

+++ src/freespace2/freespace.cc
@@ -3017,7 +3017,7 @@ game_flip_page_and_time_it()
     t2 = timer_get_fixed_seconds();
     d = t2 - t1;
     t = (gr_screen.max_w * gr_screen.max_h * gr_screen.bytes_per_pixel) / 1024;
-    sprintf(transfer_text, NOX("%d MB/s"), fixmuldiv(t, 65, d));
+    sprintf(transfer_text, NOX("%d MB/s"), (int)fixmuldiv(t, 65, d));
 }
 
 void
